Move ModeEnum into commons package
