+++ src/main/java/seedu/address/logic/commands/mode/SwitchCommand.java
@@ -1,7 +1,7 @@
 package seedu.address.logic.commands.mode;
 
 import seedu.address.logic.commands.CommandResult;
-import seedu.address.ui.ModeEnum;
+import seedu.address.commons.ModeEnum;
 
 public class SwitchCommand extends Command {
