+++ src/main/java/seedu/address/logic/parser/mode/SwitchCommandParser.java
@@ -5,7 +5,7 @@ import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT
 import seedu.address.logic.commands.mode.SwitchCommand;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.ui.ModeEnum;
+import seedu.address.commons.ModeEnum;
 
 /**
  * Parses input arguments and creates a new SwitchCommand object
