+++ src/test/java/seedu/address/logic/parser/mode/ModeParserTest.java
@@ -12,7 +12,7 @@ import seedu.address.logic.commands.mode.HelpCommand;
 import seedu.address.logic.commands.mode.SwitchCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.testutil.ModeUtil;
-import seedu.address.ui.ModeEnum;
+import seedu.address.commons.ModeEnum;
 
 public class ModeParserTest {
