+++ src/test/java/seedu/address/logic/parser/mode/SwitchCommandParserTest.java
@@ -7,7 +7,7 @@ import static seedu.address.logic.parser.mode.CommandParserTestUtil.assertParseS
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.mode.SwitchCommand;
-import seedu.address.ui.ModeEnum;
+import seedu.address.commons.ModeEnum;
 
 class SwitchCommandParserTest {
